checkKiloSortOutputs is complete!
The script checkKiloSortOutputs is the first step in creating a modular function structure - the goal of this branch. It takes outputs from KiloSort and Phy and plots rasters and SDF line plots. in the middle are packaged functions like photo diode triggering, condition selection, and plotting functions. More work needs to be done to remove extraneous manual inputs, but this is a good start.